wip: refactor Identifier Stages
