+++ src/ui/components/CreateIdentifier/CreateIdentifier.scss
@@ -1,6 +1,27 @@
 ion-modal.create-identifier-modal {
-  &.blur .create-identifier-modal-content > *:not(.spinner-container) {
-    filter: blur(0.1rem);
+  &.blur {
+    .create-identifier-modal-content > *:not(.spinner-container),
+    ion-toolbar.page-footer {
+      filter: blur(0.1rem);
+    }
+
+    ion-toolbar.page-footer {
+      background: transparent !important;
+    }
+  }
+
+  .spinner-container {
+    z-index: 1000;
+    position: absolute;
+    top: 0;
+    left: 0;
+    display: flex;
+    height: 100%;
+    width: 100%;
+    align-items: center;
+    justify-content: center;
+    border-radius: 0.5rem 0.5rem 0rem 0rem;
+    background: rgba(3, 3, 33, 0.2);
   }
 
   &:not(.full-page-modal) ion-toolbar {
@@ -27,20 +48,6 @@ ion-modal.create-identifier-modal {
   }
 
   .create-identifier-modal-content {
-    .spinner-container {
-      z-index: 100;
-      position: absolute;
-      top: 0;
-      left: 0;
-      display: flex;
-      height: 100%;
-      width: 100%;
-      align-items: center;
-      justify-content: center;
-      border-radius: 0.5rem 0.5rem 0rem 0rem;
-      background: rgba(3, 3, 33, 0.2);
-    }
-
     ion-grid {
       padding: 0;
     }
