+++ src/ui/components/CreateIdentifier/CreateIdentifier.types.ts
@@ -1,3 +1,5 @@
+import { ConnectionShortDetails } from "../../pages/Connections/Connections.types";
+
 interface CreateIdentifierProps {
   modalIsOpen: boolean;
   setModalIsOpen: (value: boolean) => void;
@@ -20,9 +22,26 @@ interface ThemeItemProps {
   index: number;
 }
 
+interface IdentifierStageProps {
+  state: {
+    multiSigStage: number;
+    displayNameValue: string;
+    selectedIdentifierType: number;
+    selectedAidType: number;
+    selectedTheme: number;
+    threshold: number;
+    selectedConnections: ConnectionShortDetails[];
+  };
+  setState: (value: any) => void;
+  componentId: string;
+  setBlur?: (value: boolean) => void;
+  resetModal: () => void;
+}
+
 export type {
   CreateIdentifierProps,
   TypeItemProps,
   IdentifierThemeSelectorProps,
   ThemeItemProps,
+  IdentifierStageProps,
 };
